refactor: Remove duplicated colors.rs file
